Implement dependency injection for SSE Service
