+++ src/Modules/Station/Components/ChatBox/ChatBox.tsx
@@ -6,7 +6,7 @@ import { Message } from 'Models';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { StationServices } from 'Services/Http';
-import { StationChatSSE } from 'Services/SSE';
+import { StationChatSSEService } from 'Services/SSE';
 import './ChatBox.scss';
 import { ChatMessage } from './ChatMessage';
 
@@ -29,7 +29,7 @@ export class ChatBoxComponent extends BaseComponent<
 > {
   private messageBox: any;
   @Inject('StationServices') private stationServices: StationServices;
-  private stationChatSSE: StationChatSSE;
+  @Inject('StationChatSSEService')private stationChatSSEService: StationChatSSEService;
 
   constructor(props: IChatBoxProps & IChatReducerProps) {
     super(props);
@@ -56,7 +56,7 @@ export class ChatBoxComponent extends BaseComponent<
   }
 
   public componentWillUnmount() {
-    this.stationChatSSE.close();
+    this.stationChatSSEService.close();
   }
 
   public componentWillReceiveProps(
@@ -94,7 +94,7 @@ export class ChatBoxComponent extends BaseComponent<
   public onSwitchStation(oldStationId: string, newStationId: string) {
     if (oldStationId !== newStationId) {
       this.setState({ listMessages: [] });
-      this.stationChatSSE.close();
+      this.stationChatSSEService.close();
       this.startSSEService(newStationId);
     }
   }
@@ -173,8 +173,8 @@ export class ChatBoxComponent extends BaseComponent<
   }
 
   private startSSEService(stationId: string) {
-    this.stationChatSSE = new StationChatSSE(stationId);
-    this.stationChatSSE.start();
+    this.stationChatSSEService.initiate(stationId);
+    this.stationChatSSEService.start();
   }
 }
