+++ src/Services/SSE/SSE.ts
@@ -1,2 +1,2 @@
-export * from './StationPlaylistSSE';
-export * from './StationChatSSE';
+export * from './StationPlaylistSSEService';
+export * from './StationChatSSEService';
